changing s3 connector to be cloud storage generic and implemented GCS and BigQuery support
