+++ example-connect-s3-sink.properties
@@ -1,5 +1,5 @@
 name=s3-sink
-connector.class=com.deviantart.kafka_connect_s3.S3SinkConnector
+connector.class=com.personali.kafka.connect.cloud.storage.CloudStorageSinkConnector
 tasks.max=1
 topics=test
 s3.bucket=connect-test
