+++ run-integration-tests.sh
@@ -21,7 +21,7 @@ sleep 10
 
 # Submitting JSON configuration
 curl -H "Content-Type: application/json" \
---data "@src/test/resources/test-connector.json" \
+--data "@src/test/resources/test-connector-S3.json" \
 http://localhost:8083/connectors
 
 # Running integration tests
