--- src/main/java/com/deviantart/kafka_connect_s3/S3SinkConnectorConstants.java
@@ -1,5 +1,0 @@
-package com.deviantart.kafka_connect_s3;
-
-public class S3SinkConnectorConstants {
-  public static final String VERSION = "0.0.1";
-}
