+++ src/main/java/com/personali/kafka/connect/cloud/storage/CloudStorageSinkConnectorConstants.java
@@ -1,0 +1,5 @@
+package com.personali.kafka.connect.cloud.storage;
+
+public class CloudStorageSinkConnectorConstants {
+  public static final String VERSION = "0.0.1";
+}
