+++ src/test/resources/test-connector-GCS.json
@@ -1,0 +1,19 @@
+{
+  "name": "cloud-storage-sink-connector",
+  "config": {
+    "connector.class": "com.personali.kafka.connect.cloud.storage.CloudStorageSinkConnector",
+    "tasks.max": "1",
+    "topics": "test-topic",
+    "storage.bucket": "personali-test",
+    "storage.prefix.test-topic": "systest",
+    "local.buffer.dir": "/tmp",
+    "custom.date.partition.format":"yyyy/MM/dd",
+    "custom.date.field.format":"yyyy-MM-dd HH:mm:ss",
+    "custom.date.field":"event_time",
+    "storage.class": "com.personali.kafka.connect.cloud.storage.storage.GCSStorageWriter",
+    "storage.partition.class":"com.personali.kafka.connect.cloud.storage.partition.CustomDateFormatStoragePartition",
+    "additional.flush.task.class": "com.personali.kafka.connect.cloud.storage.flush.FlushAdditionalBigQueryRegisterTask",
+    "bigquery.table.name": "remote-login-1121.personli.event_files_merge_manager",
+    "bigquery.register.test-topic": "true"
+  }
+}
