--- src/test/resources/test-connector.json
@@ -1,17 +1,0 @@
-{
-  "name": "s3-sink-connector",
-  "config": {
-    "connector.class": "com.deviantart.kafka_connect_s3.S3SinkConnector",
-    "tasks.max": "1",
-    "topics": "test-topic",
-    "s3.bucket": "connect-system-test",
-    "s3.prefix.test-topic": "systest",
-    "s3.endpoint": "http://fakes3:4569",
-    "s3.path_style": "true",
-    "local.buffer.dir": "/tmp",
-    "custom.date.partition.format":"yyyy/MM/dd",
-    "custom.date.field.format":"yyyy-MM-dd HH:mm:ss",
-    "custom.date.field":"event_time",
-    "s3.partition.class":"com.deviantart.kafka_connect_s3.partition.CustomDateFormatS3Partition"
-  }
-}
